feat: added functions that can be used in Rust code without the executable
* Added a function that downloads input and a function that submits answers.

Closes #12

+++ src/domain/ports/input_cache.rs
@@ -16,7 +16,7 @@ pub enum InputCacheError {
 }
 
 pub trait InputCache {
-    fn save(input: &str, year: i32, day: i32) -> Result<(), InputCacheError>;
-    fn load(year: i32, day: i32) -> Result<String, InputCacheError>;
+    fn save(input: &str, year: usize, day: usize) -> Result<(), InputCacheError>;
+    fn load(year: usize, day: usize) -> Result<String, InputCacheError>;
     fn clear() -> Result<(), InputCacheError>;
 }
