+++ src/infrastructure/find_riddle_part.rs
@@ -1,5 +1,5 @@
 use crate::domain::riddle_part::RiddlePart;
 
 pub trait FindRiddlePart {
-    fn find_unsolved_part(&self, year: i32, day: i32) -> Result<RiddlePart, anyhow::Error>;
+    fn find_unsolved_part(&self, year: usize, day: usize) -> Result<RiddlePart, anyhow::Error>;
 }
